video

+++ app/src/main/java/com/fr1014/mycoludmusic/data/entity/http/wangyiyun/playlist/Track.kt
@@ -21,7 +21,7 @@ data class Track(
         val m: M,
         val mark: Long,
         val mst: Int,
-        val mv: Int,
+        val mv: Long,
         val name: String,
         val no: Int,
         val noCopyrightRcmd: Any,
